Extracted stateExpired handler to new file
