+++ plugins/hitl/src/hooks/index.ts
@@ -1,2 +1,3 @@
 export * as beforeIncomingMessage from './before-incoming-message'
 export * as beforeIncomingEvent from './before-incoming-event'
+export * as stateExpired from './state-expired'
