+++ plugins/hitl/src/index.ts
@@ -3,8 +3,6 @@ import { isBrowser } from 'browser-or-node'
 import * as actions from './actions'
 import * as hooks from './hooks'
 import * as bp from '.botpress'
-import { getTimeoutMs, isTimedOut } from './hitl-timeout'
-import { DEFAULT_TIMEOUT_TIME } from 'plugin.definition'
 
 const plugin = new bp.Plugin({
   actions: {
@@ -68,21 +66,7 @@ plugin.on.beforeIncomingEvent('*', async (props) => {
 })
 
 plugin.on.stateExpired('hitl', async (props) => {
-  props.logger.info('HITL state expired for conversation', props.state.conversationId)
-  if (!props.state.conversationId) {
-    throw new sdk.RuntimeError('The hitl state expired without an attached conversation')
-  }
-  const conversation = await props.client.listMessages({ conversationId: props.state.conversationId })
-  const timeout = props.configuration.hitlSessionTimeoutHours ?? DEFAULT_TIMEOUT_TIME
-  if (conversation.messages[0]?.createdAt && isTimedOut(conversation.messages[0]?.createdAt, timeout)) {
-    await hooks.beforeIncomingEvent.hitlStopped.handleEvent({
-      ...props,
-      data: { payload: { conversationId: props.state.conversationId } },
-    })
-  } else {
-    const newTimeout = getTimeoutMs(timeout)
-    await props.states.conversation.hitl.set(props.state.conversationId, { hitlActive: true }, { expiryMs: newTimeout })
-  }
+  await hooks.stateExpired.hitlStateExpired(props)
 })
 
 export default plugin
